Move prrecorder and achievements into the same package
