+++ experimentation/tools/tests/_constants.py
@@ -1,0 +1,5 @@
+"""Constants for the tests."""
+import pathlib
+
+TEST_DIR = pathlib.Path(__file__).parent
+RESOURCES_DIR = TEST_DIR / "resources"
