+++ experimentation/tools/tests/test_achievements.py
@@ -2,11 +2,11 @@ import json
 import shlex
 import pathlib
 
-import achievements
+from sorald import achievements
 
-CUR_DIR = pathlib.Path(__file__).parent
-PR_RECORDER_RESOURCES = CUR_DIR.parent / "pr_recorder" / "resources"
-PRS_JSON_FINAL = PR_RECORDER_RESOURCES / "prs_final.json"
+import _constants
+
+PRS_JSON_FINAL = _constants.RESOURCES_DIR / "prs_final.json"
 
 
 def test_correctly_renders_merged_pr(tmp_path):
@@ -63,7 +63,7 @@ def test_correctly_renders_pr_without_repair_data(tmp_path):
 # Achievements
 This document presents an overview of the pull requests performed with Sorald.
 
-## [redhat-developer/rsp-server#619](https://github.com/redhat-developer/rsp-server/ulls/619)
+## [redhat-developer/rsp-server#619](https://github.com/redhat-developer/rsp-server/pull/619)
 This PR was opened at 2020-11-25 12:01:06 and merged at 2020-11-30 20:45:38.
 The patch was generated fully automatically with Sorald.
